build_and_update_model_server/requirements.txt
numpy==1.16.2
scikit-learn[alldeps]
pandas==0.25.0
Flask==1.1.1
requests==2.22.0
redis==3.3.5
imbalanced-learn==0.5.0
pyzmq==18.1.0